requirements.txt
enum34
PyYAML
docker-py
requests
